Only remove photon map(s) related to job

+++ src/util/rad.c
@@ -672,9 +672,9 @@ mkpmap(void)			/* run mkpmap if indicated */
 		if (runcom(combuf)) {
 			fprintf(stderr, "%s: error running %s\n",
 					progname, c_mkpmap);
-			if (pgmapname != NULL)
+			if (pgmapname != NULL && pgmapdate < oct1date)
 				unlink(pgmapname);
-			if (pcmapname != NULL)
+			if (pcmapname != NULL && pcmapdate < oct1date)
 				unlink(pcmapname);
 			quit(1);
 		}
